+++ doc/source/doxygen-docs/changelog.md
@@ -11,6 +11,7 @@
   - \ref mrpt_obs_grp
     - New mrpt::obs::CActionCollection::insert() overload for smart pointers.
     - New method mrpt::obs::CObservation2DRangeScan::getScanAngle() and clarify docs on class members.
+    - New class mrpt::obs::CObservation3DScene.
 
 # Version 2.3.0: Released April 25th, 2021
 - General build changes: